Refactor: 클래스네임 변경
bean 충돌이 생겨서 클래스 이름을 좀 변경 했습니다
