+++ src/main/kotlin/org/tenten/bittakotlin/feedInteraction/feedLike/dto/FeedLikeDTO.kt
@@ -1,0 +1,7 @@
+package org.tenten.bittakotlin.feedInteraction.feedLike.dto
+
+data class FeedLikeDTO(
+    val feedId: Long?,
+    val profileId: Long?,
+    val isLiked: Boolean
+)
