+++ src/main/kotlin/org/tenten/bittakotlin/feedInteraction/feedLike/service/FeedLikeService.kt
@@ -1,0 +1,8 @@
+package org.tenten.bittakotlin.feedInteraction.feedLike.service
+
+import org.tenten.bittakotlin.feedInteraction.feedLike.dto.FeedLikeDTO
+
+interface FeedLikeService {
+    fun toggleLike(feedId: Long, profileId: Long): FeedLikeDTO
+    fun getLikeCount(feedId: Long): Long
+}
