+++ src/main/kotlin/org/tenten/bittakotlin/feedInteraction/feedLike/service/FeedLikeServiceImpl.kt
@@ -1,31 +1,31 @@
-package org.tenten.bittakotlin.feedInteraction.like.service
+package org.tenten.bittakotlin.feedInteraction.feedLike.service
 
 import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.tenten.bittakotlin.feed.repository.FeedRepository
-import org.tenten.bittakotlin.feedInteraction.like.dto.LikeDTO
-import org.tenten.bittakotlin.feedInteraction.like.entity.Like
-import org.tenten.bittakotlin.feedInteraction.like.repository.LikeRepository
+import org.tenten.bittakotlin.feedInteraction.feedLike.dto.FeedLikeDTO
+import org.tenten.bittakotlin.feedInteraction.feedLike.entity.FeedLike
+import org.tenten.bittakotlin.feedInteraction.feedLike.repository.FeedLikeRepository
 import org.tenten.bittakotlin.profile.repository.ProfileRepository
 
 
 @Service
-class LikeServiceImpl(
-    private val likeRepository: LikeRepository,
+class FeedLikeServiceImpl(
+    private val likeRepository: FeedLikeRepository,
     private val feedRepository: FeedRepository,
     private val profileRepository: ProfileRepository
-) : LikeService {
+) : FeedLikeService {
 
     @Transactional
-    override fun toggleLike(feedId: Long, profileId: Long): LikeDTO {
+    override fun toggleLike(feedId: Long, profileId: Long): FeedLikeDTO {
         val feed = feedRepository.findById(feedId)
             .orElseThrow { EntityNotFoundException("Feed not found for id: $feedId") }
         val profile = profileRepository.findById(profileId)
             .orElseThrow { EntityNotFoundException("Profile not found for id: $profileId") }
 
         val like = likeRepository.findByFeedAndProfile(feed, profile).orElseGet {
-            val newLike = Like(feed = feed, profile = profile, liked = true)
+            val newLike = FeedLike(feed = feed, profile = profile, liked = true)
             likeRepository.save(newLike)
             newLike
         }
@@ -33,7 +33,7 @@ class LikeServiceImpl(
         like.liked = !like.liked
         likeRepository.save(like)
 
-        return LikeDTO(feed.id, profile.id, like.liked)
+        return FeedLikeDTO(feed.id, profile.id, like.liked)
     }
 
     @Transactional(readOnly = true)
