--- src/main/kotlin/org/tenten/bittakotlin/feedInteraction/like/dto/LikeDTO.kt
@@ -1,7 +1,0 @@
-package org.tenten.bittakotlin.feedInteraction.like.dto
-
-data class LikeDTO(
-    val feedId: Long?,
-    val profileId: Long?,
-    val isLiked: Boolean
-)
